requirements.txt
djangorestframework
decorator
PyJWT
Django==2.0.8
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
django-jwt-auth
